Move profile builder helpers into research_ai
