+++ src/research_ai/services/author_context.py
