+++ src/research_ai/services/researcher_external_context.py
